Changing OHLC data from Integer to Float and fixing datetime => date comparison bug

+++ bullwatcher.api/app/database/models.py
@@ -16,8 +16,8 @@ class StockMetadata(db.Model):
 class StockDaily(db.Model):
     date            = db.Column(db.Integer, primary_key=True)
     ticker          = db.Column(db.String(10), primary_key=True)
-    low_price       = db.Column(db.Integer)
-    high_price      = db.Column(db.Integer)
-    open_price      = db.Column(db.Integer)
-    close_price     = db.Column(db.Integer)
+    low_price       = db.Column(db.Float)
+    high_price      = db.Column(db.Float)
+    open_price      = db.Column(db.Float)
+    close_price     = db.Column(db.Float)
     volume          = db.Column(db.Integer)
